fix: fix on information banners and .
