+++ packages/twenty-front/src/modules/billing/constants/BillingCheckoutSessionDefaultValue.ts
@@ -1,0 +1,11 @@
+import { BillingCheckoutSession } from '@/auth/types/billingCheckoutSession.type';
+import {
+    BillingPlanKey,
+    SubscriptionInterval,
+} from '~/generated-metadata/graphql';
+
+export const BILLING_CHECKOUT_SESSION_DEFAULT_VALUE: BillingCheckoutSession = {
+  plan: BillingPlanKey.Pro,
+  interval: SubscriptionInterval.Month,
+  requirePaymentMethod: true,
+};
