+++ packages/twenty-front/src/modules/information-banner/components/billing/InformationBannerBillingSubscriptionPaused.tsx
@@ -1,25 +1,15 @@
 import { InformationBanner } from '@/information-banner/components/InformationBanner';
 import { AppPath } from '@/types/AppPath';
 import { SettingsPath } from '@/types/SettingsPath';
-import { useSubscriptionStatus } from '@/workspace/hooks/useSubscriptionStatus';
 import { isDefined } from 'twenty-ui';
-import {
-  SubscriptionStatus,
-  useBillingPortalSessionQuery,
-} from '~/generated/graphql';
+import { useBillingPortalSessionQuery } from '~/generated/graphql';
 
 export const InformationBannerBillingSubscriptionPaused = () => {
-  const subscriptionStatus = useSubscriptionStatus();
-
   const { data, loading } = useBillingPortalSessionQuery({
     variables: {
       returnUrlPath: `${AppPath.Settings}/${SettingsPath.Billing}`,
     },
   });
-
-  if (subscriptionStatus !== SubscriptionStatus.Paused) {
-    return null;
-  }
 
   const openBillingPortal = () => {
     if (isDefined(data) && isDefined(data.billingPortalSession.url)) {
