+++ packages/twenty-front/src/modules/information-banner/components/billing/InformationBannerFailPaymentInfo.tsx
@@ -1,28 +1,15 @@
 import { InformationBanner } from '@/information-banner/components/InformationBanner';
 import { AppPath } from '@/types/AppPath';
 import { SettingsPath } from '@/types/SettingsPath';
-import { useSubscriptionStatus } from '@/workspace/hooks/useSubscriptionStatus';
 import { isDefined } from 'twenty-ui';
-import {
-  SubscriptionStatus,
-  useBillingPortalSessionQuery,
-} from '~/generated/graphql';
+import { useBillingPortalSessionQuery } from '~/generated/graphql';
 
-export const InformationBannerBillingSubscriptionPaused = () => {
-  const subscriptionStatus = useSubscriptionStatus();
-
+export const InformationBannerFailPaymentInfo = () => {
   const { data, loading } = useBillingPortalSessionQuery({
     variables: {
       returnUrlPath: `${AppPath.Settings}/${SettingsPath.Billing}`,
     },
   });
-
-  if (
-    subscriptionStatus !== SubscriptionStatus.PastDue &&
-    subscriptionStatus !== SubscriptionStatus.Unpaid
-  ) {
-    return null;
-  }
 
   const openBillingPortal = () => {
     if (isDefined(data) && isDefined(data.billingPortalSession.url)) {
